+++ blocklets/ai-studio/api/src/routes/project.ts
@@ -1,5 +1,5 @@
 import fs from 'fs';
-import { cp, mkdtemp, readFile, rm } from 'fs/promises';
+import { mkdir, mkdtemp, readFile, rm } from 'fs/promises';
 import { basename, dirname, isAbsolute, join } from 'path';
 
 import { Config } from '@api/libs/env';
@@ -13,12 +13,14 @@ import {
   isAssistant,
   nextAssistantId,
   projectSettingsSchema,
+  variableToYjs,
 } from '@blocklet/ai-runtime/types';
+import { copyRecursive } from '@blocklet/ai-runtime/utils/fs';
 import { call } from '@blocklet/sdk/lib/component';
 import config from '@blocklet/sdk/lib/config';
 import { user } from '@blocklet/sdk/lib/middlewares';
 import { Request, Router } from 'express';
-import { exists, pathExists } from 'fs-extra';
+import { exists } from 'fs-extra';
 import * as git from 'isomorphic-git';
 import http from 'isomorphic-git/http/node';
 import Joi from 'joi';
@@ -37,16 +39,17 @@ import {
 } from '../libs/security';
 import Project, { nextProjectId } from '../store/models/project';
 import {
+  ASSETS_DIR,
   CONFIG_FILENAME,
   CONFIG_FILE_KEY,
   CONFIG_FOLDER,
   LOGO_FILENAME,
+  ProjectRepo,
+  SETTINGS_FILE,
   VARIABLE_FILENAME,
   VARIABLE_KEY,
   autoSyncIfNeeded,
   clearRepository,
-  commitProjectSettingWorking,
-  commitWorking,
   defaultBranch,
   defaultRemote,
   getAssistantFromRepository,
@@ -111,7 +114,6 @@ export interface UpdateProjectInput {
   name?: string;
   description?: string;
   pinned?: boolean;
-  icon?: string;
   model?: string;
   temperature?: number;
   topP?: number;
@@ -138,7 +140,6 @@ const updateProjectSchema = Joi.object<UpdateProjectInput>({
   name: Joi.string().empty([null, '']),
   description: Joi.string().empty([null, '']),
   pinned: Joi.boolean().empty([null]),
-  icon: Joi.string().allow('').empty([null, '']),
   model: Joi.string().empty([null, '']),
   temperature: Joi.number().min(0).max(2).empty(null),
   topP: Joi.number().min(0.1).max(1).empty(null),
@@ -269,11 +270,20 @@ export function projectRoutes(router: Router) {
       list.map(async (project) => {
         const repository = await getRepository({ projectId: project.id });
         const branches = await repository.listBranches();
+
+        const working = await repository.working({ ref: project.gitDefaultBranch });
+        const projectSetting = working.syncedStore.files[SETTINGS_FILE] as ProjectSettings | undefined;
+
         const users = await getAuthorsOfProject({
           projectId: project.id,
           gitDefaultBranch: project.gitDefaultBranch,
         });
-        return { ...project.dataValues, users, branches };
+        return {
+          ...project.dataValues,
+          users,
+          branches,
+          ...pick(projectSetting, 'iconVersion', 'name', 'description'),
+        };
       })
     );
 
@@ -334,28 +344,50 @@ export function projectRoutes(router: Router) {
     res.json({ icons: data?.uploads || [] });
   });
 
+  const logoQuerySchema = Joi.object<{ blockletDid?: string; projectRef?: string; working?: boolean }>({
+    blockletDid: Joi.string().empty(['', null]),
+    projectRef: Joi.string().empty(['', null]),
+    working: Joi.boolean().empty(['', null]),
+  });
+
   router.get('/projects/:projectId/logo.png', async (req, res) => {
     const { projectId } = req.params;
     if (!projectId) throw new Error('Missing required parameter `projectId`');
 
-    const original = await Project.findOne({ where: { id: projectId } });
-    if (original) {
-      const repository = await getRepository({ projectId });
-      const { blob } = await repository.readBlob({ ref: original.gitDefaultBranch, filepath: LOGO_FILENAME });
-      res.setHeader('Content-Type', 'image/png');
-      res.end(Buffer.from(blob));
-      return;
-    }
+    const query = await logoQuerySchema.validateAsync(req.query, { stripUnknown: true });
 
-    const resource = await getProjectFromResource({ projectId });
+    if (query.blockletDid) {
+      const resource = await getProjectFromResource({ blockletDid: query.blockletDid, projectId });
 
-    if (resource) {
       if (resource?.gitLogoPath && (await exists(resource.gitLogoPath))) {
         res.sendFile(resource.gitLogoPath);
         return;
       }
 
       throw new NotFoundError(`No such project icon ${projectId}`);
+    }
+
+    const original = await Project.findOne({ where: { id: projectId } });
+    if (original) {
+      const repository = await getRepository({ projectId });
+      const working = await repository.working({ ref: query.projectRef || original.gitDefaultBranch });
+
+      if (query.working) {
+        const logoPath = join(working.workingDir, LOGO_FILENAME);
+        if (await exists(logoPath)) {
+          res.setHeader('Content-Type', 'image/png');
+          res.sendFile(logoPath);
+          return;
+        }
+      }
+
+      const { blob } = await repository.readBlob({
+        ref: query.projectRef || original.gitDefaultBranch,
+        filepath: LOGO_FILENAME,
+      });
+      res.setHeader('Content-Type', 'image/png');
+      res.end(Buffer.from(blob));
+      return;
     }
 
     throw new NotFoundError(`No such project ${projectId}`);
@@ -632,23 +664,9 @@ export function projectRoutes(router: Router) {
 
     checkProjectPermission({ req, project });
 
-    const {
-      name,
-      pinned,
-      description,
-      icon,
-      model,
-      temperature,
-      topP,
-      presencePenalty,
-      frequencyPenalty,
-      maxTokens,
-      gitType,
-      gitAutoSync,
-      didSpaceAutoSync,
-      homePageUrl,
-      appearance,
-    } = await updateProjectSchema.validateAsync(req.body, { stripUnknown: true });
+    const { pinned, gitType, gitAutoSync, didSpaceAutoSync } = await updateProjectSchema.validateAsync(req.body, {
+      stripUnknown: true,
+    });
 
     if (gitAutoSync) {
       const repo = await getRepository({ projectId });
@@ -664,22 +682,11 @@ export function projectRoutes(router: Router) {
     await project.update(
       omitBy(
         {
-          name,
           pinnedAt: pinned ? new Date().toISOString() : pinned === false ? null : undefined,
           updatedBy: userId,
-          description,
-          model: model || project.model || '',
-          icon: '',
-          temperature,
-          topP,
-          presencePenalty,
-          frequencyPenalty,
-          maxTokens,
           gitType,
           gitAutoSync,
           didSpaceAutoSync,
-          homePageUrl,
-          appearance,
           updatedAt: new Date(),
         },
         (v) => v === undefined
@@ -687,8 +694,6 @@ export function projectRoutes(router: Router) {
     );
 
     const author = { name: fullName, email: userId };
-
-    await commitProjectSettingWorking({ project, author, icon });
 
     await autoSyncIfNeeded({ project, author, userId });
 
@@ -921,6 +926,9 @@ export function projectRoutes(router: Router) {
 
       const repo = await getRepository({ projectId });
 
+      const working = await repo.working({ ref });
+      const settings = working.syncedStore.files[SETTINGS_FILE];
+
       const agent = await getAssistantFromRepository({
         repository: repo,
         ref,
@@ -928,7 +936,7 @@ export function projectRoutes(router: Router) {
         working: query.working,
       });
 
-      res.json({ agent, project: project.dataValues });
+      res.json({ agent, project: { ...project.dataValues, ...(working ? settings : {}) } });
     }
   );
 
@@ -999,6 +1007,49 @@ export function projectRoutes(router: Router) {
 
     return res.json({ assistants: uniqBy(assistants, 'id') });
   });
+
+  const uploadAssetSchema = Joi.object<{ type: 'logo' | 'asset'; source: string }>({
+    type: Joi.string().valid('logo', 'asset').empty(['', null]).default('asset'),
+    source: Joi.string().required(),
+  });
+
+  router.post(
+    '/projects/:projectId/refs/:ref/assets',
+    user(),
+    ensureComponentCallOrPromptsEditor(),
+    async (req, res) => {
+      const { projectId, ref } = req.params;
+      if (!projectId || !ref) throw new Error('Missing required params `projectId` or `ref`');
+
+      const input = await uploadAssetSchema.validateAsync(req.body, { stripUnknown: true });
+
+      const project = await Project.findByPk(projectId, { rejectOnEmpty: new Error('Project not found') });
+
+      checkProjectPermission({ req, project });
+
+      const repo = await ProjectRepo.load({ projectId });
+      const { filename, hash } = await repo.uploadAsset({ type: input.type, ref, source: input.source });
+
+      res.json({ filename, hash });
+    }
+  );
+
+  router.get('/projects/:projectId/refs/:ref/assets/:filename', async (req, res) => {
+    const { projectId, ref, filename } = req.params;
+    if (!projectId || !ref || !filename) throw new Error('Missing required params `projectId` or `ref` or `filename`');
+
+    await Project.findByPk(projectId, { rejectOnEmpty: new Error('Project not found') });
+
+    const repo = await ProjectRepo.load({ projectId });
+    const working = await repo.working({ ref });
+
+    const p = join(working.workingDir, ASSETS_DIR, filename);
+    if (await exists(p)) {
+      return res.sendFile(p);
+    }
+
+    return res.status(404).end();
+  });
 }
 
 const getAuthorsOfProject = async ({
@@ -1045,9 +1096,9 @@ async function copyProject({
   });
 
   const parent = dirname(repo.root);
-  await cp(repo.root, join(parent, project.id!), { recursive: true });
-  if (await pathExists(`${repo.root}.cooperative`)) {
-    await cp(`${repo.root}.cooperative`, join(parent, `${project.id}.cooperative`), { recursive: true });
+  await copyRecursive(repo.root, join(parent, project.id!));
+  if (await exists(`${repo.root}.cooperative`)) {
+    await copyRecursive(`${repo.root}.cooperative`, join(parent, `${project.id}.cooperative`));
   }
 
   return project;
@@ -1056,6 +1107,8 @@ async function copyProject({
 async function createProjectFromTemplate(
   template: Omit<(typeof projectTemplates)[number], 'project'> & {
     project: Omit<(typeof projectTemplates)[number]['project'], 'createdAt' | 'updatedAt'>;
+    assetsDir?: string;
+    gitLogoPath?: string;
   },
   {
     name,
@@ -1087,6 +1140,7 @@ async function createProjectFromTemplate(
   });
 
   const working = await repository.working({ ref: defaultBranch });
+
   for (const { parent, ...file } of template.assistants) {
     const id = file.id || nextAssistantId();
     const assistant = fileToYjs({ ...file, id });
@@ -1108,17 +1162,25 @@ async function createProjectFromTemplate(
   }
 
   working.syncedStore.tree[VARIABLE_KEY] = joinURL(CONFIG_FOLDER, VARIABLE_FILENAME);
-  working.syncedStore.files[VARIABLE_KEY] = { type: 'variables', variables: template.memory?.variables ?? [] };
+  working.syncedStore.files[VARIABLE_KEY] = {
+    type: 'variables',
+    variables: (template.memory?.variables ?? []).map(variableToYjs),
+  };
   working.syncedStore.tree[CONFIG_FILE_KEY] = joinURL(CONFIG_FOLDER, CONFIG_FILENAME);
   working.syncedStore.files[CONFIG_FILE_KEY] = template.config || {};
 
-  await commitWorking({
-    project,
+  const assetsDir = join(working.workingDir, 'assets/');
+  await mkdir(assetsDir, { recursive: true });
+  if (template.assetsDir && (await exists(template.assetsDir))) {
+    await copyRecursive(join(template.assetsDir, '/'), assetsDir);
+  }
+
+  await repository.commitWorking({
     ref: defaultBranch,
     branch: defaultBranch,
     message: 'First Commit',
     author: { name: author.fullName, email: author.did },
-    icon: (template as any)?.gitLogoPath || (await sampleIcon()),
+    icon: template?.gitLogoPath || (await sampleIcon()),
   });
 
   return project;